refactor: move legacy proximity utils and types to storybook
